# Fix NPC performance issues and optimize lighting