… scenes/npc.tscn